+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Performance
 
 - Avoid map lookups in some `properties` validators.
+- Faster validation of `{"type": "array", "items": {...}}` schemas.
 
 ## [0.48.2] - 2026-07-21